refactor(repository) : 레포지토리 디렉토리 변경 #14
