+++ src/main/java/com/my/relink/controller/trade/dto/request/TradeReqDto.java
@@ -1,0 +1,8 @@
+package com.my.relink.controller.trade.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class TradeReqDto {
+    Long userId;
+}
