+++ src/main/java/com/my/relink/domain/point/pointHistory/repository/PointHistoryRepository.java
@@ -4,7 +4,6 @@ import com.my.relink.domain.point.pointHistory.PointHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
 import java.util.Optional;
 
 @Repository
