--- src/main/java/com/my/relink/domain/trade/dto/TradeRequestDto.java
@@ -1,8 +1,0 @@
-package com.my.relink.domain.trade.dto;
-
-import lombok.Getter;
-
-@Getter
-public class TradeRequestDto {
-    Long userId;
-}
