+++ src/main/java/com/my/relink/domain/user/repository/UserRepository.java
@@ -2,8 +2,6 @@ package com.my.relink.domain.user.repository;
 
 import com.my.relink.domain.user.User;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
 
-@Repository
 public interface UserRepository extends JpaRepository<User, Long> {
 }
